refactor: split ResolveMacroPath into single and list variants
- BaseResolveMacroPath holds shared _extract_path_string and _resolve_one helpers
- ResolveMacroPath: single path in (ParameterString) → single resolved_path out
- ResolveMacroPaths: ParameterList in → resolved_paths list out
- Both are SuccessFailureNode with proper failure routing
